backend/incident_agent/requirements.txt
mysql-connector-python
PyYAML
openai
groq